+++ .travis.yml
@@ -9,13 +9,13 @@ matrix:
   - jdk: oraclejdk8
     env: TERM=dumb MOCK_MAKER=mock-maker-inline KOTLIN_VERSION=1.1.61
   - jdk: oraclejdk8
-    env: TERM=dumb MOCK_MAKER=mock-maker-inline KOTLIN_VERSION=1.2.10
+    env: TERM=dumb MOCK_MAKER=mock-maker-inline KOTLIN_VERSION=1.2.20
   - jdk: oraclejdk8
     env: TERM=dumb KOTLIN_VERSION=1.0.7
   - jdk: oraclejdk8
     env: TERM=dumb KOTLIN_VERSION=1.1.61
   - jdk: oraclejdk8
-    env: TERM=dumb KOTLIN_VERSION=1.2.10
+    env: TERM=dumb KOTLIN_VERSION=1.2.20
 
 
 env:
@@ -26,7 +26,10 @@ install:
  - true
 
 script:
-  - ./gradlew test
+  - ops/mockMakerInline.sh
+  - ./gradlew clean
+  - ./gradlew assemble
+  - ./gradlew test -i
 
 notifications:
   email: false
