--- mockito-kotlin/src/main/kotlin/com/nhaarman/mockitokotlin2/internal/MockMaker.kt
@@ -1,35 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright (c) 2018 Niek Haarman
- * Copyright (c) 2007 Mockito contributors
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-
-package com.nhaarman.mockitokotlin2.internal
-
-import org.mockito.internal.configuration.plugins.Plugins
-import org.mockito.internal.creation.bytebuddy.InlineByteBuddyMockMaker
-
-internal var mockMakerInlineEnabled: Boolean? = null
-internal fun mockMakerInlineEnabled(): Boolean {
-    return mockMakerInlineEnabled ?:
-          Plugins.getMockMaker() is InlineByteBuddyMockMaker
-}
