--- mockito-kotlin/src/test/kotlin/test/TestBase.kt
@@ -1,18 +1,0 @@
-package test
-
-import com.nhaarman.mockitokotlin2.internal.mockMakerInlineEnabled
-import org.junit.After
-import org.junit.Before
-
-abstract class TestBase {
-
-    @Before
-    open fun setup() {
-        mockMakerInlineEnabled = false
-    }
-
-    @After
-    open fun tearDown() {
-        mockMakerInlineEnabled = null
-    }
-}
