+++ settings.gradle
@@ -1,1 +1,2 @@
 include 'mockito-kotlin'
+include 'tests'
