+++ tests/build.gradle
@@ -1,0 +1,29 @@
+buildscript {
+    ext.kotlin_version = '1.2.20'
+    ext.kotlin_version = System.getenv("KOTLIN_VERSION") ?: '1.0.7'
+
+    repositories {
+        mavenCentral()
+    }
+    dependencies {
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+    }
+}
+
+apply plugin: 'java'
+apply plugin: 'kotlin'
+
+repositories {
+    mavenCentral()
+    jcenter()
+}
+
+dependencies {
+    compile files("${rootProject.projectDir}/mockito-kotlin/build/libs/mockito-kotlin-${rootProject.ext.versionName}.jar")
+
+    compile "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    compile "org.mockito:mockito-core:2.13.0"
+
+    testCompile "junit:junit:4.12"
+    testCompile "com.nhaarman:expect.kt:1.0.0"
+}
