+++ tests/src/test/kotlin/test/TestBase.kt
@@ -1,0 +1,11 @@
+package test
+
+import org.junit.After
+
+abstract class TestBase {
+
+    @After
+    open fun tearDown() {
+        mockMakerInlineEnabled = null
+    }
+}
