+++ tests/src/test/kotlin/test/inline/UsingMockMakerInlineTest.kt
@@ -23,15 +23,11 @@
  */
 
 import com.nhaarman.expect.expect
-import com.nhaarman.mockitokotlin2.any
-import com.nhaarman.mockitokotlin2.internal.mockMakerInlineEnabled
-import com.nhaarman.mockitokotlin2.doReturn
-import com.nhaarman.mockitokotlin2.mock
-import com.nhaarman.mockitokotlin2.verify
-import com.nhaarman.mockitokotlin2.whenever
+import com.nhaarman.mockitokotlin2.*
 import org.junit.Assume.assumeTrue
 import org.junit.Before
 import org.junit.Test
+import test.mockMakerInlineEnabled
 import java.io.IOException
 import java.math.BigInteger
